Add files via upload
My research internship work under Dr. Bidyut Patra, IIT(BHU), Varanasi.
This is the outcome of a research internship conducted in the field of machine learning, with a specific focus on recommender systems. Recommender systems play a crucial role in various domains by providing personalized recommendations based on user preferences and historical data. This research internship aimed to explore and develop innovative algorithms and methodologies to generate accurate and effective recommendations using the widely-used Movie Lens dataset.
Throughout the internship, practical experience was gained in implementing machine learning algorithms. This included data preprocessing, feature engineering, algorithm development, and evaluation techniques. The research process involved problem formulation, experimental design, and result analysis to address the challenges associated with building effective recommender systems.
The outcomes of this research internship have the potential to enhance user experiences and improve the quality of recommendations. The findings contribute to the broader field of machine learning and recommender systems, advancing the understanding and development of personalized recommendation technologies.
In conclusion, the research internship in machine learning under the recommender system has been a valuable and rewarding experience. The project has provided insights into the development and implementation of recommender systems, specifically using the MovieLens dataset. By exploring various machine learning algorithms and methodologies, we have successfully developed a recommender system that generates accurate and personalized recommendations. The outcomes of this internship have the potential to enhance user experiences, improve recommendation quality, and contribute to the broader field of machine learning and recommender systems. The knowledge and practical experience gained during this internship will serve as a foundation for future research and advancements in personalized recommendation technologies.